requirements.txt
nltk
sentence-transformers
scikit-learn
python-docx
pandas
openpyxl
transformers
openai
python-dotenv
pdfminer.six